// garethwhite/src/types/storyblok-component-types.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from '@storyblok/js';
import type { StoryblokAsset, StoryblokRichtext } from '../../.storyblok/types/storyblok';
export interface Article {
  title?: string;
  components?: (Picture | RichtextBlock)[];
  feature_image?: StoryblokAsset;
  component: "article";
  _uid: string;
  [k: string]: unknown;
}

export interface ArticleListing {
  title?: string;
  sort_order?: "" | "desc" | "asc";
  component: "article_listing";
  _uid: string;
  [k: string]: unknown;
}

export interface Feature {
  name?: string;
  component: "feature";
  _uid: string;
  [k: string]: unknown;
}

export interface FrontPage {
  title?: string;
  features?: (ISbStoryData<Article> | string)[];
  component: "front_page";
  _uid: string;
  [k: string]: unknown;
}

export interface Page {
  title?: string;
  body?: (Article | ArticleListing | Feature | FrontPage | Page | Picture | RichtextBlock)[];
  component: "page";
  _uid: string;
  [k: string]: unknown;
}

export interface Picture {
  image?: StoryblokAsset;
  component: "picture";
  _uid: string;
  [k: string]: unknown;
}

export interface RichtextBlock {
  title?: string;
  richtext?: StoryblokRichtext;
  component: "richtext_block";
  _uid: string;
  [k: string]: unknown;
}

export type ContentType = Article | ArticleListing | FrontPage | Page;
// Aliases for use in component props (import from "@/types/storyblok-component-types")
export type ArticleStoryblok = Article;
export type ArticleListingStoryblok = ArticleListing;
export type FeatureStoryblok = Feature;
export type FrontPageStoryblok = FrontPage;
export type PageStoryblok = Page;
export type PictureStoryblok = Picture;
export type RichtextBlockStoryblok = RichtextBlock;